refactor(api): removes useless service
